requirements.txt
streamlit>=1.20
ultralytics>=8.0.0
transformers>=4.30.0
sentencepiece
Pillow
pyttsx3
playsound==1.2.2
numpy
opencv-python
huggingface-hub
roboflow